requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.9.0,<3.11.0
djangorestframework-simplejwt>=5.0.0,<5.1.0
Pillow>=9.0.0